got existing tests passing with ReportModel
