+++ TestStack.BDDfy.Tests/Reporters/Html/MetroReportBuilderTests.cs
@@ -12,7 +12,7 @@ namespace TestStack.BDDfy.Tests.Reporters.Html
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void ShouldProduceExpectedHtml()
         {
-            Func<FileReportModel> model = () => 
+            var model = 
                 new HtmlReportModel(new ReportTestData().CreateMixContainingEachTypeOfOutcome().ToReportModel())
                     {
                         RunDate = new DateTime(2014, 3, 25, 11, 30, 5)
@@ -26,7 +26,7 @@ namespace TestStack.BDDfy.Tests.Reporters.Html
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void ShouldProduceExpectedHtmlWithExamples()
         {
-            Func<FileReportModel> model = () =>
+            var model = 
                 new HtmlReportModel(new ReportTestData().CreateTwoStoriesEachWithOneFailingScenarioAndOnePassingScenarioWithThreeStepsOfFiveMillisecondsAndEachHasTwoExamples()
                     .ToReportModel())
                 {
