+++ TestStack.BDDfy.Tests/Reporters/MarkDown/MarkDownReportBuilderTests.cs
@@ -12,7 +12,7 @@ namespace TestStack.BDDfy.Tests.Reporters.MarkDown
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void ShouldProduceExpectedMarkdown()
         {
-            Func<FileReportModel> model = () => new FileReportModel(new ReportTestData().CreateMixContainingEachTypeOfOutcome().ToReportModel());
+            var model = new FileReportModel(new ReportTestData().CreateMixContainingEachTypeOfOutcome().ToReportModel());
             var sut = new MarkDownReportBuilder();
             ReportApprover.Approve(model, sut);
         }
@@ -21,7 +21,7 @@ namespace TestStack.BDDfy.Tests.Reporters.MarkDown
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void ShouldProduceExpectedMarkdownWithExamples()
         {
-            Func<FileReportModel> model = () =>
+            var model = 
                 new FileReportModel(new ReportTestData().CreateTwoStoriesEachWithOneFailingScenarioAndOnePassingScenarioWithThreeStepsOfFiveMillisecondsAndEachHasTwoExamples()
                     .ToReportModel())
                 {
