+++ TestStack.BDDfy.Tests/Reporters/ReportApprover.cs
@@ -9,12 +9,12 @@ namespace TestStack.BDDfy.Tests.Reporters
 {
     class ReportApprover
     {
-        public static void Approve(Func<FileReportModel> model, IReportBuilder reportBuilder)
+        public static void Approve(FileReportModel model, IReportBuilder reportBuilder)
         {
             // setting the culture to make sure the date is formatted the same on all machines
             using (new TemporaryCulture("en-GB"))
             {
-                var result = reportBuilder.CreateReport(model());
+                var result = reportBuilder.CreateReport(model);
                 Approvals.Verify(result, s => Scrub(StackTraceScrubber.ScrubLineNumbers(StackTraceScrubber.ScrubPaths(s))));
             }
         }
