+++ TestStack.BDDfy.Tests/TagsTests.cs
@@ -29,13 +29,9 @@ namespace TestStack.BDDfy.Tests
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void TagsAreReportedInHtmlReport()
         {
-            var story = this.Given(_ => GivenAStep())
-                .WithTags("Tag1", "Tag 2")
-                .BDDfy();
-            Func<FileReportModel> model = () => new HtmlReportModel(new[] { story }.ToReportModel())
-                {
-                    RunDate = new DateTime(2014, 3, 25, 11, 30, 5)
-                };
+            var model = new HtmlReportModel(this.CreateReportModel())            {
+                RunDate = new DateTime(2014, 3, 25, 11, 30, 5)
+            };
 
             var sut = new ClassicReportBuilder();
             ReportApprover.Approve(model, sut);
@@ -45,13 +41,11 @@ namespace TestStack.BDDfy.Tests
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void TagsAreReportedInMetroHtmlReport()
         {
-            var story = this.Given(_ => GivenAStep())
-                .WithTags("Tag1", "Tag 2")
-                .BDDfy();
-            Func<FileReportModel> model = () => new HtmlReportModel(new[] { story }.ToReportModel())
-                {
-                    RunDate = new DateTime(2014, 3, 25, 11, 30, 5)
-                };
+            var reportModel = this.CreateReportModel();
+            var model = new HtmlReportModel(reportModel)
+            {
+                RunDate = new DateTime(2014, 3, 25, 11, 30, 5)
+            };
 
             var sut = new MetroReportBuilder();
             ReportApprover.Approve(model, sut);
@@ -61,10 +55,8 @@ namespace TestStack.BDDfy.Tests
         [MethodImpl(MethodImplOptions.NoInlining)]
         public void TagsAreReportedInMarkdownReport()
         {
-            var story = this.Given(_ => GivenAStep())
-                .WithTags("Tag1", "Tag 2")
-                .BDDfy();
-            Func<FileReportModel> model = () => new FileReportModel(new[] { story }.ToReportModel());
+            var reportModel = this.CreateReportModel();
+            var model = new FileReportModel(reportModel);
             var sut = new MarkDownReportBuilder();
             ReportApprover.Approve(model, sut);
         }
@@ -72,6 +64,15 @@ namespace TestStack.BDDfy.Tests
         private void GivenAStep()
         {
 
+        }
+
+        private ReportModel CreateReportModel()
+        {
+            var story = this.Given(_ => GivenAStep())
+                .WithTags("Tag1", "Tag 2")
+                .BDDfy();
+            var reportModel = new[] { story }.ToReportModel();
+            return reportModel;
         }
     }
 }
